access boundary conditions in advection schemes and specify boundary conditions as a dictionary to allow for future developments
