+++ gusto/__init__.py
@@ -5,8 +5,8 @@ from gusto.diffusion import *                        # noqa
 from gusto.eady_diagnostics import *                 # noqa
 from gusto.equations import *                        # noqa
 from gusto.forcing import *                          # noqa
-from gusto.form_manipulation_labelling import *      # noqa
 from gusto.initialisation_tools import *             # noqa
+from gusto.labels import *                           # noqa
 from gusto.limiters import *                         # noqa
 from gusto.linear_solvers import *                   # noqa
 from gusto.physics import *                          # noqa
@@ -15,3 +15,4 @@ from gusto.recovery import *                         # noqa
 from gusto.state import *                            # noqa
 from gusto.timeloop import *                         # noqa
 from gusto.transport_equation import *               # noqa
+from gusto.fml import *                              # noqa
